Yields

Yields.AbstractYieldCurve

Yields.ForwardYields

Yields.AbstractYieldCurve — Type

An AbstractInterestCurve is an object which can be called with:

rate for the spot rate at a given time

disc for the spot discount rate at a given time

Yields.ForwardYields — Method

Forwards(rate_vector)

Takes a vector of 1-period forward rates and constructs a discount curve.