FinanceModels.Volatility API Reference

FinanceModels.Volatility.Constant

Exported API

Unexported API

FinanceModels.Volatility.Constant — Type

Volatility.Constant(σ)

A constant volatility per period. If σ is not explicitly passed, then it is set to zero.

Please open an issue if you encounter any issues or confusion with the package.