Yields API Reference

Please open an issue if you encounter any issues or confusion with the package.

Rate Types

When JuliaActuary packages return a rate, they will be of a Rate type, such as Rate(0.05,Periodic(2)) for a 5% rate compounded twice per period. It is recommended to keep rates typed and use them throughout the ecosystem without modifying it.

For example, if we construct a curve like this:

# 2021-03-31 rates from Treasury.gov
rates =[0.01, 0.01, 0.03, 0.05, 0.07, 0.16, 0.35, 0.92, 1.40, 1.74, 2.31, 2.41] ./ 100
mats = [1/12, 2/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30]
  
curve = Yields.CMT(rates,mats)

Then rates from this curve will be typed. For example:

z = zero(c,10)

Now, z will be: Yields.Rate{Float64, Continuous}(0.01779624378877313, Continuous())

This Rate has both the rate an the compounding convention embedded in the datatype.

You can now use that rate throughout the JuliaActuary ecosystem, such as with ActuaryUtilities.jl:

using ActuaryUtilities
present_values(z,cashflows)

If you need to extract the rate for some reason, you can get the rate by calling Yields.rate(...). Using the above example, Yields.rate(z) will return 0.01779624378877313.

Yields.Bootstrap

Yields.BulletBondQuote

Yields.Constant

Yields.ForwardStarting

Yields.NelsonSiegel

Yields.NelsonSiegelCurve

Yields.NelsonSiegelSvensson

Yields.NelsonSiegelSvenssonCurve

Yields.RateCombination

Yields.SmithWilson

Yields.Step

Yields.SwapQuote

Yields.YieldCurveFitParameters

Yields.ZeroCouponQuote

Base.:*

Base.:+

Base.:-

Base.:/

Base.zero

FinanceCore.accumulation

FinanceCore.discount

FinanceCore.forward

Yields.CMT

Yields.Forward

Yields.H

Yields.H_ordered

Yields.OIS

Yields.Par

Yields.Zero

Yields.bootstrap

Yields.cashflows

Yields.par

Yields.timepoints

Yields.Bootstrap — Type

Boostrap(interpolation_method=QuadraticSpline)

This YieldCurveFitParameters object defines the interpolation method to use when bootstrapping the curve. Provided options are QuadraticSpline() (the default) and LinearSpline(). You may also pass a custom interpolation method with the function signature of f(xs, ys) -> f(x) -> y.

If constructing curves and the rates are not Rates (ie you pass a Vector{Float64}), then they will be interpreted as Periodic(1) Rates, except the Par curve, which is interpreted as Periodic(2) Rates. CMT and OIS FinanceCore.CompoundingFrequency assumption depends on the corresponding maturity.

See for more:

Zero

Forward

Par

CMT

OIS

Yields.BulletBondQuote — Type

BulletBondQuote(yield, price, maturity, frequency)

Quote for a set of fixed interest bullet bonds with given yield, price, maturity and a given payment frequency frequency.

Construct a vector of quotes for use with SmithWilson methods, e.g. by broadcasting over an array of inputs.

Examples

julia> maturities = [1.2, 2.5, 3.6]
julia> interests = [-0.02, 0.3, 0.04]
julia> prices = [1.3, 0.1, 4.5]
julia> frequencies = [2,1,2]
julia> bbq = Yields.BulletBondQuote.(interests, maturities, prices, frequencies)

Yields.Constant — Type

Constant(rate::Real, cf::CompoundingFrequency=Periodic(1))
Constant(r::Rate)

Construct a yield object where the spot rate is constant for all maturities. If rate is not a Rate type, will assume Periodic(1) for the compounding frequency

Examples

julia> y = Yields.Constant(0.05)
julia> FinanceCore.discount(y,2)
0.9070294784580498     # 1 / (1.05) ^ 2

Yields.ForwardStarting — Type

ForwardStarting(curve,forwardstart)

Rebase a curve so that discount/accumulation/etc. are re-based so that time zero from the new curves perspective is the given forwardstart time.

Examples

julia> zero = [5.0, 5.8, 6.4, 6.8] ./ 100
julia> maturity = [0.5, 1.0, 1.5, 2.0]
julia> curve = Yields.Zero(zero, maturity)
julia> fwd = Yields.ForwardStarting(curve, 1.0)

julia> FinanceCore.discount(curve,1,2)
0.9275624570410582

julia> FinanceCore.discount(fwd,1) # `curve` has effectively been reindexed to `1.0`
0.9275624570410582

Extended Help

While ForwardStarting could be nested so that, e.g. the third period's curve is the one-period forward of the second period's curve, it will be more efficient to reuse the initial curve from a runtime and compiler perspective.

ForwardStarting is not used to construct a curve based on forward rates. See Forward instead.

Yields.NelsonSiegel — Type

NelsonSiegel(τ_initial)
NelsonSiegel() # defaults to τ_initial=1.0

This parameter set is used to fit the Nelson-Siegel parametric model to given rates. τ_initial should be a scalar and is used as the starting τ value in the optimization. The default value for τ_initial is 1.0.

When fitting rates using this YieldCurveFitParameters object, the Nelson-Siegel model is used. If constructing curves and the rates are not Rates (ie you pass a Vector{Float64}), then they will be interpreted as Continuous Rates.

See for more:

Zero

Forward

Par

CMT

OIS

Yields.NelsonSiegelCurve — Type

NelsonSiegel(rates::AbstractVector, maturities::AbstractVector; τ_initial=1.0)

Return the NelsonSiegel fitted parameters. The rates should be zero spot rates. If rates are not Rates, then they will be interpreted as Continuous Rates.

NelsonSiegel(β₀, β₁, β₂, τ₁)

Parameters of Nelson and Siegel (1987) parametric model:

β₀ represents a long-term interest rate

β₁ represents a time-decay component

β₂ represents a hump

τ₁ controls the location of the hump

Examples

julia> β₀, β₁, β₂, τ₁ = 0.6, -1.2, -1.9, 3.0
julia> nsm = Yields.NelsonSiegel.(β₀, β₁, β₂, τ₁)

# Extend Help

## References
- https://onriskandreturn.com/2019/12/01/nelson-siegel-yield-curve-model/
- https://www.bis.org/publ/bppdf/bispap25.pdf

Yields.NelsonSiegelSvensson — Type

NelsonSiegelSvensson(τ_initial) 
NelsonSiegelSvensson() # defaults to τ_initial=[1.0,1.0]

This parameter set is used to fit the Nelson-Siegel parametric model to given rates. τ_initial should be a two element vector and is used as the starting τ value in the optimization. The default value for τ_initial is [1.0,1.0].

See for more:

Zero

Forward

Par

CMT

OIS

Yields.NelsonSiegelSvenssonCurve — Type

NelsonSiegelSvensson(yields::AbstractVector, maturities::AbstractVector; τ_initial=[1.0,1.0])

Return the NelsonSiegelSvensson fitted parameters. The rates should be continuous zero spot rates. If rates are not Rates, then they will be interpreted as Continuous Rates.

When fitting rates using this YieldCurveFitParameters object, the Nelson-Siegel model is used. If constructing curves and the rates are not Rates (ie you pass a Vector{Float64}), then they will be interpreted as Continuous Rates.

See for more:

- [`Zero`](@ref)
- [`Forward`](@ref)
- [`Par`](@ref)
- [`CMT`](@ref)
- [`OIS`](@ref)

NelsonSiegelSvensson(β₀, β₁, β₂, β₃, τ₁, τ₂)

Parameters of Svensson (1994) parametric model:

β₀ represents a long-term interest rate

β₁ represents a time-decay component

β₂ represents a hump

β₃ represents a second hum

τ₁ controls the location of the hump

τ₁ controls the location of the second hump

Examples

julia> β₀, β₁, β₂, β₃, τ₁, τ₂ = 0.6, -1.2, -2.1, 3.0, 1.5
julia> nssm = NelsonSiegelSvensson.NelsonSiegelSvensson.(β₀, β₁, β₂, β₃, τ₁, τ₂)

## References
- https://onriskandreturn.com/2019/12/01/nelson-siegel-yield-curve-model/
- https://www.bis.org/publ/bppdf/bispap25.pdf

Yields.RateCombination — Type

RateCombination(curve1,curve2,operation)

Creates a datastructure that will perform the given operation after independently calculating the effects of the two curves. Can only be created via the public API by using the +, -, *, and / operatations on AbstractYield objects.

As this is double the normal operations when performing calculations, if you are using the curve in performance critical locations, you should consider transforming the inputs and constructing a single curve object ahead of time.

Yields.SmithWilson — Type

SmithWilson(zcq::Vector{ZeroCouponQuote}; ufr, α)
SmithWilson(swq::Vector{SwapQuote}; ufr, α)
SmithWilson(bbq::Vector{BulletBondQuote}; ufr, α)
SmithWilson(times<:AbstractVector, cashflows<:AbstractMatrix, prices<:AbstractVector; ufr, α)
SmithWilson(u, qb; ufr, α)

Create a yield curve object that implements the Smith-Wilson interpolation/extrapolation scheme.

Positional arguments to construct a curve:

Quoted instrument as the first argument: either a Vector of ZeroCouponQuotes, SwapQuotes, or BulletBondQuotes, or

A set of times, cashflows, and prices, or

A curve can be with u is the timepoints coming from the calibration, and qb is the internal parameterization of the curve that ensures that the calibration is correct. Users may prefer the other constructors but this mathematical constructor is also available.

Required keyword arguments:

ufr is the Ultimate Forward Rate, the forward interest rate to which the yield curve tends, in continuous compounding convention.

α is the parameter that governs the speed of convergence towards the Ultimate Forward Rate. It can be typed with \alpha[TAB]

Yields.Step — Type

Step(rates,times)

Create a yield curve object where the applicable rate is the effective rate of interest applicable until corresponding time. If rates is not a Vector{Rate}, will assume Periodic(1) type.

The last rate will be applied to any time after the last time in times.

Examples

julia>y = Yields.Step([0.02,0.05], [1,2])

julia>rate(y,0.5)
0.02

julia>rate(y,1.5)
0.05

julia>rate(y,2.5)
0.05

Yields.SwapQuote — Type

SwapQuote(yield, maturity, frequency)

Quote for a set of interest rate swaps with the given yield and maturity and a given payment frequency.

Examples

julia> maturities = [1.2, 2.5, 3.6]
julia> interests = [-0.02, 0.3, 0.04]
julia> prices = [1.3, 0.1, 4.5]
julia> frequencies = [2,1,2]
julia> swq = Yields.SwapQuote.(interests, maturities, frequencies)

Yields.YieldCurveFitParameters — Type

A YieldCurveFitParameters is a structure which contains associated parameters for a yield curve fitting procedure. The type of the object determines the method, and the values of the object determine the parameters.

If the fitting data and the rates are passed as <:Real numbers instead of a type of Rates, the default interpretation may vary depending on the fitting type/parameter. See the individual docstrings of the types for more information.

Available types are:

Bootstrap

NelsonSiegel

NelsonSiegelSvensson

Yields.ZeroCouponQuote — Type

ZeroCouponQuote(price, maturity)

Quote for a set of zero coupon bonds with given price and maturity.

Examples

julia> prices = [1.3, 0.1, 4.5]
julia> maturities = [1.2, 2.5, 3.6]
julia> swq = Yields.ZeroCouponQuote.(prices, maturities)

Base.:* — Method

Yields.AbstractYieldCurve * Yields.AbstractYieldCurve

The multiplication of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the two curves will be added together. This can be useful, for example, if you wanted to after-tax a yield.

Examples

julia> m = Yields.Constant(0.01) * 0.79;

julia> accumulation(m,1)
1.0079

julia> accumulation(.01*.79,1)
1.0079

Base.:+ — Method

Yields.AbstractYieldCurve + Yields.AbstractYieldCurve

The addition of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the two curves will be added together.

Base.:- — Method

Yields.AbstractYieldCurve - Yields.AbstractYieldCurve

The subtraction of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the second curves will be subtracted from the first.

Base.:/ — Method

Yields.AbstractYieldCurve / Yields.AbstractYieldCurve

The division of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the two curves will have the first divided by the second. This can be useful, for example, if you wanted to gross-up a yield to be pre-tax.

Examples

julia> m = Yields.Constant(0.01) / 0.79;

julia> accumulation(d,1)
1.0126582278481013

julia> accumulation(.01/.79,1)
1.0126582278481013

Base.zero — Method

zero(curve,time)
zero(curve,time,CompoundingFrequency)

Return the zero rate for the curve at the given time.

FinanceCore.accumulation — Method

accumulation(yc, from, to)

The accumulation factor for the yield curve yc for times from through to.

FinanceCore.discount — Method

discount(yc, to)
discount(yc, from,to)

The discount factor for the yield curve yc for times from through to.

FinanceCore.forward — Method

forward(yc, from, to, CompoundingFrequency=Periodic(1))

The forward Rate implied by the yield curve yc between times from and to.

Yields.CMT — Function

Yields.CMT(rates, maturities; interpolation=QuadraticSpline())
Yields.CMT(p::YieldCurveFitParameters,rates, maturities)

Takes constant maturity (treasury) yields (bond equivalent), and assumes that instruments <= one year maturity pay no coupons and that the rest pay semi-annual.

The method of fitting the curve to the data is determined by the YieldCurveFitParameters object p, which is a Boostrap(QuadraticSpline()) by default.

Examples

# 2021-03-31 rates from Treasury.gov
rates =[0.01, 0.01, 0.03, 0.05, 0.07, 0.16, 0.35, 0.92, 1.40, 1.74, 2.31, 2.41] ./ 100
mats = [1/12, 2/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30]
	
Yields.CMT(rates,mats)

Yields.Forward — Function

Forward(rates,maturities=eachindex(rates))
Forward(p::YieldCurveFitParameters, rates,maturities=eachindex(rates))

Takes a vector of 1-period forward rates and constructs a discount curve. The method of fitting the curve to the data is determined by the YieldCurveFitParameters object p, which is a Boostrap(QuadraticSpline()) by default.

If rates is a vector of floating point number instead of a vector Rates, see the YieldCurveFitParameters for how the rate will be interpreted.

Examples

julia> Yields.Forward( [0.01,0.02,0.03] );

julia> Yields.Forward( Yields.Continuous.([0.01,0.02,0.03]) );

Yields.H — Method

H(α, t1, t2)

The Smith-Wilson H function implemented in a faster way.

Yields.H_ordered — Method

H_ordered(α, t_min, t_max)

The Smith-Wilson H function with ordered arguments (for better performance than using min and max).

Yields.OIS — Function

OIS(rates, maturities)
OIS(p::YieldCurveFitParameters, rates, maturities)

Takes Overnight Index Swap rates, and assumes that instruments <= one year maturity are settled once and other agreements are settled quarterly with a corresponding CompoundingFrequency.

The method of fitting the curve to the data is determined by the YieldCurveFitParameters object p, which is a Boostrap(QuadraticSpline()) by default.

Examples

julia> ois = [1.8, 2.0, 2.2, 2.5, 3.0, 4.0] ./ 100;
julia> mats = [1 / 12, 1 / 4, 1 / 2, 1, 2, 5];
julia> curve = Yields.OIS(ois, mats)

              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Yield Curve (Yields.BootstrapCurve)⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           
              ┌────────────────────────────────────────────────────────────┐           
          0.1 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ Zero rates
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
   Continuous │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⡤⠔⠒⠋⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⡠⠔⠊⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⣀⠔⠋⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⡠⠎⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠜⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
         0.01 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              └────────────────────────────────────────────────────────────┘           
              ⠀0⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀time⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀30⠀   

Yields.Par — Function

Par(rates, maturities=eachindex(rates)
Par(p::YieldCurveFitParameters, rates, maturities=eachindex(rates)

Construct a curve given a set of bond equivalent yields and the corresponding maturities. Assumes that maturities <= 1 year do not pay coupons and that after one year, pays coupons with frequency equal to the CompoundingFrequency of the corresponding rate (normally the default for a Rate is 1, but when constructed via Par the default compounding Frequency is 2).

See bootstrap for more on the interpolation parameter, which is set to QuadraticSpline() by default.

Examples

julia> par = [6.,8.,9.5,10.5,11.0,11.25,11.38,11.44,11.48,11.5] ./ 100
julia> maturities = [t for t in 1:10]
julia> curve = Par(par,maturities);
julia> zero(curve,1)
Rate(0.06000000000000005, Periodic(1))

Yields.Zero — Function

Zero(rates, maturities=eachindex(rates))
Zero(p::YieldCurveFitParameters,rates, maturities=eachindex(rates))

Construct a yield curve with given zero-coupon spot rates at the given maturities. The method of fitting the curve to the data is determined by the YieldCurveFitParameters object p, which is a Boostrap(QuadraticSpline()) by default.

If rates is a vector of floating point number instead of a vector Rates, see the YieldCurveFitParameters for how the rate will be interpreted.

Examples

julia> Yields.Zero([0.01,0.02,0.04,0.05],[1,2,5,10])

              ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀Yield Curve (Yields.BootstrapCurve)⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀           
              ┌────────────────────────────────────────────────────────────┐           
         0.05 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡠⠤⠒⠒⠒⠒⠒⠤⠤⠤⢄⣀⣀⣀⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│ Zero rates
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⠖⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠉⠉⠒⠒⠒⠢⠤⠤⠤⣄⣀⣀⣀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⠖⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠉⠑⠒⠒⠒⠦⠤⠤⠤⣀⣀⣀⡀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⡔⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠉│           
              │⠀⠀⠀⠀⠀⠀⠀⢠⠎⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⢠⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
   Continuous │⠀⠀⠀⠀⢀⠇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⡜⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⡸⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⢠⠇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⡜⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠒⠒⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
            0 │⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀│           
              └────────────────────────────────────────────────────────────┘           
              ⠀0⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀time⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀30⠀      

Yields.bootstrap — Function

bootstrap(rates, maturities, settlement_frequency, interpolation::QuadraticSpline())

Bootstrap the rates with the given maturities, treating the rates according to the periodic frequencies in settlement_frequency.

interpolator is any function that will take two vectors of inputs and output points and return a function that will estimate an output given a scalar input. That is interpolator should be: interpolator(xs, ys) -> f(x) where f(x) is the interpolated value of y at x.

Built in interpolators in Yields are:

QuadraticSpline(): Quadratic spline interpolation.

LinearSpline(): Linear spline interpolation.

The default is QuadraticSpline().

Yields.cashflows — Method

cashflows(interests, maturities, frequency)
timepoints(zcq::Vector{ZeroCouponQuote})
timepoints(bbq::Vector{BulletBondQuote})

Produce a cash flow matrix for a set of instruments with given interests and maturities and a given payment frequency frequency. All instruments are assumed to have their first payment at time 1/frequency and have their last payment at the largest multiple of 1/frequency less than or equal to the input maturity.

Yields.par — Method

par(curve,time;frequency=2)

Calculate the par yield for maturity time for the given curve and frequency. Returns a Rate object with periodicity corresponding to the frequency. The exception to this is if time is less than what the payments allowed by frequency (e.g. a time 0.5 but with frequency 1) will effectively assume frequency equal to 1 over time.

Examples

julia> c = Yields.Constant(0.04);

julia> Yields.par(c,4)
Yields.Rate{Float64, Yields.Periodic}(0.03960780543711406, Yields.Periodic(2))

julia> Yields.par(c,4;frequency=1)
Yields.Rate{Float64, Yields.Periodic}(0.040000000000000036, Yields.Periodic(1))

julia> Yields.par(c,0.6;frequency=4)
Yields.Rate{Float64, Yields.Periodic}(0.039413626195875295, Yields.Periodic(4))

julia> Yields.par(c,0.2;frequency=4)
Yields.Rate{Float64, Yields.Periodic}(0.039374942589460726, Yields.Periodic(5))

julia> Yields.par(c,2.5)
Yields.Rate{Float64, Yields.Periodic}(0.03960780543711406, Yields.Periodic(2))

Yields.timepoints — Method

timepoints(zcq::Vector{ZeroCouponQuote})
timepoints(bbq::Vector{BulletBondQuote})

Return the times associated with the cashflows of the instruments.