Yields

Yields.AbstractYield

Yields.Forward

Yields.Par

Yields.AbstractYield — Type

An AbstractInterestCurve is an object which can be called with:

rate for the spot rate at a given time

discount for the spot discount rate at a given time

Yields.Forward — Method

Forward(rate_vector)

Takes a vector of 1-period forward rates and constructs a discount curve.

Yields.Par — Method

Construct a curve given a set of bond yields priced at par with a single coupon per period.