FinanceModels.Spline API Reference

FinanceModels.Spline.Cubic

FinanceModels.Spline.Linear

FinanceModels.Spline.Quadratic

Exported API

FinanceModels.Spline — Module

Spline is a module which offers various degree splines used for fitting or bootstraping curves via the fit function.

Available methods:

Spline.BSpline(n) where n is the nth order. A spline function of order n is a piecewise polynomial function of degree n − 1. This means that, e.g., cubic polynomial is a fourth degree B-Spline.

This object is not a fitted spline itself, rather it is a placeholder object which will be a spline representing the data only after using within fit.

and convienience methods which create a Spline.BSpline object of the appropriate order.

Spline.Linear()

Spline.Quadratic()

Spline.Cubic()

Unexported API

FinanceModels.Spline.Cubic — Method

Spline.Cubic()

Create a cubic B-spline. This object is not a fitted spline itself, rather it is a placeholder object which will be a spline representing the data only after using within fit.

Returns

A BSpline object representing a cubic B-spline.

Examples

julia> Spline.Cubic()
BSpline(4)

FinanceModels.Spline.Linear — Method

Spline.Linear()

Create a linear B-spline. This object is not a fitted spline itself, rather it is a placeholder object which will be a spline representing the data only after using within fit.

Returns

A BSpline object representing a linear B-spline.

Examples

julia> Spline.Linear()
BSpline(2)

FinanceModels.Spline.Quadratic — Method

Spline.Quadratic()

Create a quadratic B-spline. This object is not a fitted spline itself, rather it is a placeholder object which will be a spline representing the data only after using within fit.

Returns

A BSpline object representing a quadratic B-spline.

Examples

julia> Spline.Quadratic()
BSpline(3)

Please open an issue if you encounter any issues or confusion with the package.