Yields API Reference

Please open an issue if you encounter any issues or confusion with the package.

Yields.AbstractYield

Yields.Constant

Yields.Continuous

Yields.Continuous

Yields.Periodic

Yields.Periodic

Yields.Rate

Yields.Step

Base.:+

Base.:-

Base.convert

Base.zero

Yields.CMT

Yields.Forward

Yields.OIS

Yields.Par

Yields.Zero

Yields.accumulation

Yields.discount

Yields.AbstractYield — Type

An AbstractYield is an object which can be used as an argument to:

zero-coupon spot rates viea zero

discount factor via discount

accumulation factor via accumulation

It can be be constructed via:

zero rate curve with Zero

forward rate curve with Forward

par rate curve with Par

typical OIS curve with OIS

typical constant maturity treasury (CMT) curve with CMT

Yields.Constant — Type

Constant(rate)

Construct a yield object where the spot rate is constant for all maturities. If rate is not a Rate type, will assume Periodic(1) for the compounding frequency

Examples

julia> y = Yields.Constant(0.05)
julia> discount(y,2)
0.9070294784580498     # 1 / (1.05) ^ 2

Yields.Continuous — Type

Continuous()

A type representing continuous interest compounding frequency.

Examples

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

See also: Periodic

Yields.Continuous — Method

Continuous(rate)

A convinience constructor for Rate(x,Continuous())

julia> Continuous(0.01)
Rate(0.01, Continuous())

See also: Periodic

Yields.Periodic — Type

Periodic(frequency)

A type representing periodic interest compounding with the given frequency

Examples

Creating a semi-annual bond equivalent yield:

julia> Rate(0.01,Periodic(2))
Rate(0.01, Periodic(2))

See also: Continuous

Yields.Periodic — Method

Periodic(rate,frequency)

A convinience constructor for Rate(rate,Periodic(frequency)).

Examples

Creating a semi-annual bond equivalent yield:

julia> Periodic(0.01,2)
Rate(0.01, Periodic(2))

See also: Continuous

Yields.Rate — Method

Rate(rate[,frequency=1])
Rate(rate,frequency::CompoundingFrequency)

Rate is a type that encapsulates an interest rate along wtih its compounding frequency.

Periodic rates can be constructed via Rate(rate,frequency) or Rate(rate,Periodic(frequency)).

Continuous rates can be constructed via Rate(rate, Inf) or Rate(rate,Continuous()).

Examples

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

julia> Rate(0.01,Periodic(2))
Rate(0.01, Periodic(2))

julia> Rate(0.01)
Rate(0.01, Periodic(1))

julia> Rate(0.01,2)
Rate(0.01, Periodic(2))

julia> Rate(0.01,Periodic(4))
Rate(0.01, Periodic(4))

julia> Rate(0.01,Inf)
Rate(0.01, Continuous())

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

Yields.Step — Type

Step(rates,times)

Create a yield curve object where the applicable rate is the effective rate of interest applicable until corresponding time. If rates is not a Vector{Rate}, will assume Periodic(1) type.

Examples

julia>y = Yields.Step([0.02,0.05], [1,2])

julia>rate(y,0.5)
0.02

julia>rate(y,1.5)
0.05

julia>rate(y,2.5)
0.05

Base.:+ — Method

Yields.AbstractYield + Yields.AbstractYield

The addition of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the two curves will be added together.

Base.:- — Method

Yields.AbstractYield - Yields.AbstractYield

The subtraction of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the second curves will be subtracted from the first.

Base.convert — Method

convert(T::CompoundingFrequency,r::Rate)

Returns a Rate with an equivalent discount but represented with a different compounding frequency.

Examples

julia> r = Rate(Periodic(12),0.01)
Rate(0.01, Periodic(12))

julia> convert(Periodic(1),r)
Rate(0.010045960887181016, Periodic(1))

julia> convert(Continuous(),r)
Rate(0.009995835646701251, Continuous())

Base.zero — Method

zero(curve,time)
zero(curve,time,CompoundingFrequency)

Return the zero rate for the curve at the given time. If not specified, will use Periodic(1) compounding.

Yields.CMT — Method

Takes CMT yields (bond equivalent), and assumes that instruments <= one year maturity pay no coupons and that the rest pay semi-annual.

Yields.Forward — Method

Forward(rate_vector,maturities)

Takes a vector of 1-period forward rates and constructs a discount curve.

Yields.OIS — Method

OIS(rates,maturities)

Takes Overnight Index Swap rates, and assumes that instruments <= one year maturity are settled once and other agreements are settled quarterly with a corresponding CompoundingFrequency

Yields.Par — Method

Par(rate, maturity)

Construct a curve given a set of bond equivalent yields and the corresponding maturities. Assumes that maturities <= 1 year do not pay coupons and that after one year, pays coupons with frequency equal to the CompoundingFrequency of the corresponding rate.

Examples

julia> par = [6.,8.,9.5,10.5,11.0,11.25,11.38,11.44,11.48,11.5] ./ 100
julia> maturities = [t for t in 1:10]
julia> curve = Par(par,maturities);
julia> zero(curve,1)
Rate(0.06000000000000005, Periodic(1))

Yields.Zero — Method

Zero(rates,maturities)

Construct a yield curve with given zero-coupon spot rates at the given maturities. If rates is not a Vector{Rate}, will assume Periodic(1) type.

Yields.accumulation — Method

accumulation(rate,from,to)

The accumulation factor for the rate for times from through to. If rate is a Real number, will assume a Constant interest rate.

Yields.discount — Method

discount(rate,to)
discount(rate,from,to)

The discount factor for the rate for times from through to. If rate is a Real number, will assume a Constant interest rate.