Yields API Reference

Please open an issue if you encounter any issues or confusion with the package.

Yields.AbstractYield

Yields.BulletBondQuote

Yields.Constant

Yields.Continuous

Yields.Continuous

Yields.Periodic

Yields.Periodic

Yields.Rate

Yields.SmithWilson

Yields.Step

Yields.SwapQuote

Yields.ZeroCouponQuote

Base.:+

Base.:-

Base.convert

Base.zero

Yields.CMT

Yields.Forward

Yields.H

Yields.H_ordered

Yields.OIS

Yields.Par

Yields.Zero

Yields.accumulation

Yields.cashflows

Yields.discount

Yields.forward

Yields.timepoints

Yields.AbstractYield — Type

An AbstractYield is an object which can be used as an argument to:

zero-coupon spot rates via zero

discount factor via discount

accumulation factor via accumulation

It can be be constructed via:

zero rate curve with Zero

forward rate curve with Forward

par rate curve with Par

typical OIS curve with OIS

typical constant maturity treasury (CMT) curve with CMT

Yields.BulletBondQuote — Type

BulletBondQuote(yield, price, maturity, frequency)

Quote for a set of fixed interest bullet bonds with given yield, price, maturity and a given payment frequency frequency.

Construct a vector of quotes for use with SmithWilson methods, e.g. by broadcasting over an array of inputs.

Examples

julia> maturities = [1.2, 2.5, 3.6]
julia> interests = [-0.02, 0.3, 0.04]
julia> prices = [1.3, 0.1, 4.5]
julia> frequencies = [2,1,2]
julia> bbq = Yields.BulletBondQuote.(interests, maturities, prices, frequencies)

Yields.Constant — Type

Constant(rate)

Construct a yield object where the spot rate is constant for all maturities. If rate is not a Rate type, will assume Periodic(1) for the compounding frequency

Examples

julia> y = Yields.Constant(0.05)
julia> discount(y,2)
0.9070294784580498     # 1 / (1.05) ^ 2

Yields.Continuous — Type

Continuous()

A type representing continuous interest compounding frequency.

Examples

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

See also: Periodic

Yields.Continuous — Method

julia> Continuous(0.01)
Rate(0.01, Continuous())

See also: Periodic

Yields.Periodic — Type

Periodic(frequency)

A type representing periodic interest compounding with the given frequency

Examples

Creating a semi-annual bond equivalent yield:

julia> Rate(0.01,Periodic(2))
Rate(0.01, Periodic(2))

See also: Continuous

Yields.Periodic — Method

Periodic(rate,frequency)

A convenience constructor for Rate(rate,Periodic(frequency)).

Examples

Creating a semi-annual bond equivalent yield:

julia> Periodic(0.01,2)
Rate(0.01, Periodic(2))

See also: Continuous

Yields.Rate — Method

Rate(rate[,frequency=1])
Rate(rate,frequency::CompoundingFrequency)

Rate is a type that encapsulates an interest rate along with its compounding frequency.

Periodic rates can be constructed via Rate(rate,frequency) or Rate(rate,Periodic(frequency)).

Continuous rates can be constructed via Rate(rate, Inf) or Rate(rate,Continuous()).

Examples

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

julia> Rate(0.01,Periodic(2))
Rate(0.01, Periodic(2))

julia> Rate(0.01)
Rate(0.01, Periodic(1))

julia> Rate(0.01,2)
Rate(0.01, Periodic(2))

julia> Rate(0.01,Periodic(4))
Rate(0.01, Periodic(4))

julia> Rate(0.01,Inf)
Rate(0.01, Continuous())

julia> Rate(0.01,Continuous())
Rate(0.01, Continuous())

Yields.SmithWilson — Type

SmithWilson(zcq::Vector{ZeroCouponQuote}; ufr, α)
SmithWilson(swq::Vector{SwapQuote}; ufr, α)
SmithWilson(bbq::Vector{BulletBondQuote}; ufr, α)
SmithWilson(times<:AbstractVector, cashflows<:AbstractMatrix, prices<:AbstractVector; ufr, α)
SmithWilson(u, qb; ufr, α)

Create a yield curve object that implements the Smith-Wilson interpolation/extrapolation scheme.

Positional arguments to construct a curve:

Quoted instrument as the first argument: either a Vector of ZeroCouponQuotes, SwapQuotes, or BulletBondQuotes, or

A set of times, cashflows, and prices, or

A curve can be with u is the timepoints coming from the calibration, and qb is the internal parameterization of the curve that ensures that the calibration is correct. Users may prefer the other constructors but this mathematical constructor is also available.

Required keyword arguments:

ufr is the Ultimate Forward Rate, the forward interest rate to which the yield curve tends, in continuous compounding convention.

α is the parameter that governs the speed of convergence towards the Ultimate Forward Rate. It can be typed with \alpha[TAB]

Yields.Step — Type

Step(rates,times)

Create a yield curve object where the applicable rate is the effective rate of interest applicable until corresponding time. If rates is not a Vector{Rate}, will assume Periodic(1) type.

Examples

julia>y = Yields.Step([0.02,0.05], [1,2])

julia>rate(y,0.5)
0.02

julia>rate(y,1.5)
0.05

julia>rate(y,2.5)
0.05

Yields.SwapQuote — Type

SwapQuote(yield, maturity, frequency)

Quote for a set of interest rate swaps with the given yield and maturity and a given payment frequency.

Examples

julia> maturities = [1.2, 2.5, 3.6]
julia> interests = [-0.02, 0.3, 0.04]
julia> prices = [1.3, 0.1, 4.5]
julia> frequencies = [2,1,2]
julia> swq = Yields.SwapQuote.(interests, maturities, frequencies)

Yields.ZeroCouponQuote — Type

ZeroCouponQuote(price, maturity)

Quote for a set of zero coupon bonds with given price and maturity.

Examples

julia> prices = [1.3, 0.1, 4.5]
julia> maturities = [1.2, 2.5, 3.6]
julia> swq = Yields.ZeroCouponQuote.(prices, maturities)

Base.:+ — Method

Yields.AbstractYield + Yields.AbstractYield

The addition of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the two curves will be added together.

Base.:- — Method

Yields.AbstractYield - Yields.AbstractYield

The subtraction of two yields will create a RateCombination. For rate, discount, and accumulation purposes the spot rates of the second curves will be subtracted from the first.

Base.convert — Method

convert(T::CompoundingFrequency,r::Rate)

Returns a Rate with an equivalent discount but represented with a different compounding frequency.

Examples

julia> r = Rate(Periodic(12),0.01)
Rate(0.01, Periodic(12))

julia> convert(Periodic(1),r)
Rate(0.010045960887181016, Periodic(1))

julia> convert(Continuous(),r)
Rate(0.009995835646701251, Continuous())

Base.zero — Method

zero(curve,time)
zero(curve,time,CompoundingFrequency)

Return the zero rate for the curve at the given time. If not specified, will use Periodic(1) compounding.

Yields.CMT — Method

Takes CMT yields (bond equivalent), and assumes that instruments <= one year maturity pay no coupons and that the rest pay semi-annual.

Yields.Forward — Method

Forward(rate_vector,maturities)

Takes a vector of 1-period forward rates and constructs a discount curve.

Yields.H — Method

H(α, t1, t2)

The Smith-Wilson H function implemented in a faster way.

Yields.H_ordered — Method

H_ordered(α, t_min, t_max)

The Smith-Wilson H function with ordered arguments (for better performance than using min and max).

Yields.OIS — Method

OIS(rates,maturities)

Takes Overnight Index Swap rates, and assumes that instruments <= one year maturity are settled once and other agreements are settled quarterly with a corresponding CompoundingFrequency

Yields.Par — Method

Par(rate, maturity)

Construct a curve given a set of bond equivalent yields and the corresponding maturities. Assumes that maturities <= 1 year do not pay coupons and that after one year, pays coupons with frequency equal to the CompoundingFrequency of the corresponding rate.

Examples

julia> par = [6.,8.,9.5,10.5,11.0,11.25,11.38,11.44,11.48,11.5] ./ 100
julia> maturities = [t for t in 1:10]
julia> curve = Par(par,maturities);
julia> zero(curve,1)
Rate(0.06000000000000005, Periodic(1))

Yields.Zero — Method

Zero(rates,maturities)

Construct a yield curve with given zero-coupon spot rates at the given maturities. If rates is not a Vector{Rate}, will assume Periodic(1) type.

Yields.accumulation — Method

accumulation(rate,from,to)

The accumulation factor for the rate for times from through to. If rate is a Real number, will assume a Constant interest rate.

Yields.cashflows — Method

cashflows(interests, maturities, frequency)
timepoints(zcq::Vector{ZeroCouponQuote})
timepoints(bbq::Vector{BulletBondQuote})

Produce a cash flow matrix for a set of instruments with given interests and maturities and a given payment frequency frequency. All instruments are assumed to have their first payment at time 1/frequency and have their last payment at the largest multiple of 1/frequency less than or equal to the input maturity.

Yields.discount — Method

discount(rate,to)
discount(rate,from,to)

The discount factor for the rate for times from through to. If rate is a Real number, will assume a Constant interest rate.

Yields.forward — Method

forward(curve,from,to,CompoundingFrequency=Periodic(1))

The forward Rate implied by the curve between times from and to.

Yields.timepoints — Method

timepoints(zcq::Vector{ZeroCouponQuote})
timepoints(bbq::Vector{BulletBondQuote})

Return the times associated with the cashflows of the instruments.